Update README.md to include THL
Based on the hundreds of thousands of #dataviz tweets already processed, noise is a little over 92% of the conversation. The platform filters and ranks the most popular shared content in realtime. Data Visualisation’s zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, tracking dozens of topics. It will always be free to use. No ads. No BS.