Move entry to books section; update Wikipedia links
The entry for "Data Visualisation: A Handbook for Data Driven Design" was introduced in commit 5452af5. I suspect it was added to the wrong section by mistake, so I'm moving it to the "Books" section.

In addition, some of the Wikipedia links have become outdated since I introduced them in 0e3cad6 almost two years ago, due to changes in the structure of the linked articles. This commit updates the anchors to correspond to the current section names of those articles.